backend/explainability.py
import torch
import numpy as np
import cv2
from PIL import Image

def generate_heatmap(model, image_tensor):
    """
    Placeholder for Grad-CAM implementation.
    In a real implementation, you would:
    1. Register hooks on the last convolutional layer.
    2. Run forward pass.
    3. Run backward pass for the target class.
    4. Compute weights and combine with feature maps.
    
    Since we are using a pre-trained model without specific fine-tuning for this 'demo' code,
    we rely on the mock implementation in model.py for the visual effect.
    This file serves as the architectural place where real Grad-CAM logic would live.
    """
    pass
